+++ CHANGES.rst
@@ -1,6 +1,14 @@
 Changelog
 =========
 
+Version 0.9.4
+-------------
+
+- Adding sync_dirs function
+- Adding sanitized_input function (Thanks to Dogeek)
+- Breaking change: Changing find_files to return pathlib objects by default on Python 3.4+
+- Removing python 2.6 from travis tests
+
 Version 0.9.3
 -------------